# config/swagger-api-docs.json
{
    "openapi": "3.0.0",
    "info": {
      "description": "This file to describe MPYA Project",
      "version": "1.0.0",
      "title": "MPYA API",
      "contact": {
        "email": "achrafboudawarafseg@gmail.com"
      }
    },
    "paths": {
...
      "/user/document/{id_document}/versions": {
        "get": {
          "tags": [
            "User Version"
          ],
          "summary": "User document versions",
          "parameters": [
            {
              "in": "path",
              "name": "id_document",
              "schema": {
                "type": "string"
              },
              "required": true,
              "description": "ID of the document"
            }
          ],
          "description": "",
          "operationId": "",
          "responses": {
            "200": {
              "description": "Get User document versions",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Version"
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized"
            }
          },
          "security": [
            {
              "BearerAuth": []
            }
          ]
        }
      },
      "/user/version/{id_version}": {
        "get": {
          "tags": [
            "User Version"
          ],
          "summary": "version by id",
          "parameters": [
            {
              "in": "path",
              "name": "id_version",
              "schema": {
                "type": "string"
              },
              "required": true,
              "description": "ID of the version"
            }
          ],
          "description": "",
          "operationId": "",
          "responses": {
            "200": {
              "description": "Get version",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Version"
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized"
            }
          },
          "security": [
            {
              "BearerAuth": []
            }
          ]
        }
      },
      "/user/version/{id_version}/preview": {
        "get": {
          "tags": [
            "User Version"
          ],
          "summary": "version image",
          "parameters": [
            {
              "in": "path",
              "name": "id_version",
              "schema": {
                "type": "string"
              },
              "required": true,
              "description": "ID of the version"
            }
          ],
          "description": "",
          "operationId": "",
          "responses": {
            "200": {
              "description": "version image",
              "content": {
                "image/png": {
                  "schema": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized"
            }
          },
          "security": [
            {
              "BearerAuth": []
            }
          ]
        }
      },
      "/user/version/{id_version}/file": {
        "get": {
          "tags": [
            "User Version"
          ],
          "summary": "version file",
          "parameters": [
            {
              "in": "path",
              "name": "id_version",
              "schema": {
                "type": "string"
              },
              "required": true,
              "description": "ID of the version"
            }
          ],
          "description": "",
          "operationId": "",
          "responses": {
            "200": {
              "description": "version file",
              "content": {
                "application/pdf": {
                  "schema": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized"
            }
          },
          "security": [
            {
              "BearerAuth": []
            }
          ]
        }
      },
...
    },
    "servers": [
      {
        "url": "https://localhost:3000/"
      }
    ],
    "components": {
      "securitySchemes": {
        "BearerAuth": {
          "type": "http",
          "scheme": "bearer"
        }
      },
      "schemas": {
        "User": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "readOnly": true
            },
            "firstName": {
              "type": "string"
            },
            "lastName": {
              "type": "string"
            },
            "birthDate": {
              "type": "string"
            },
            "company": {
              "type": "string"
            },
            "email": {
              "type": "string"
            },
            "phone": {
              "type": "string"
            },
            "createdAt": {
              "type": "string",
              "readOnly": true
            },
            "updatedAt": {
              "type": "string",
              "readOnly": true
            }
          },
          "xml": {
            "name": "User"
          }
        },
        "Document": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "readOnly": true
            },
            "name": {
              "type": "string"
            },
            "address": {
              "type": "string"
            },
            "latitude": {
              "type": "string"
            },
            "longitude": {
              "type": "string"
            },
            "identifier": {
              "type": "string",
              "readOnly": true
            },
            "keywords": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "events": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nbVersions": {
              "type": "integer",
              "readOnly": true
            },
            "nbViews": {
              "type": "integer",
              "readOnly": true
            },
            "nbSaves": {
              "type": "integer",
              "readOnly": true
            },
            "nbDownloads": {
              "type": "integer",
              "readOnly": true
            },
            "disabledAt": {
              "type": "string",
              "readOnly": true
            },
            "user": {
              "$ref": "#/components/schemas/User"
            },
            "createdAt": {
              "type": "string",
              "readOnly": true
            },
            "updatedAt": {
              "type": "string",
              "readOnly": true
            }
          },
          "xml": {
            "name": "Document"
          }
        },
...
        "Version": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "readOnly": true
            },
            "name": {
              "type": "string"
            },
            "number": {
              "type": "integer",
              "readOnly": true
            },
            "document": {
              "$ref": "#/components/schemas/Document"
            },
            "nbViews": {
              "type": "integer",
              "readOnly": true
            },
            "nbSaves": {
              "type": "integer",
              "readOnly": true
            },
            "nbDownloads": {
              "type": "integer",
              "readOnly": true
            },
            "createdAt": {
              "type": "string",
              "readOnly": true
            },
            "updatedAt": {
              "type": "string",
              "readOnly": true
            }
          }
        },
...
      }
    }
  }